# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ADS API.",
        "title": "ADSolutions API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1/",
    "paths": {
...
        "/modelpointvariables/{prodId}": {
            "get": {
                "description": "get all applicable model point variables for a product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of model point variables for product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product Code",
                        "name": "prodId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ModelPointVariable"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ModelPointVariable": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "POLICY_NUMBER"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Policy Number"
                },
                "type": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}